requirements.txt
PyPDF2==3.0.1
nltk==3.8.1
transformers==4.36.0
torch==2.1.0
python-docx==1.0.1
